requirements.txt
flask==2.3.3
psycopg2-binary==2.9.7
pgvector==0.2.3
sentence-transformers==2.2.2
openai==1.3.5
numpy==1.24.3
faiss-cpu==1.7.4
python-dotenv==1.0.0
PyGithub==1.59.1
requests==2.31.0

